+++ core/src/engine.rs
@@ -95,7 +95,9 @@ pub fn general_handle_event(
         }
 
         Event::Leave { player_addr } => {
-            if context
+            if !context.allow_exit {
+                Err(Error::CantLeave)
+            } else if context
                 .players
                 .iter()
                 .find(|p| p.addr.eq(player_addr))